Change RemarkCommand to throw an exception
Follow the convention and throw a CommandException instead of returning
a CommandResult as the command did not fulfill its intended function.

Add RemarkCommandTest that will test that
execute_remark_throwsCommandException() throws an Exception.

Modify AddressBookParser to accept a RemarkCommand.

Add new test to AddressBookParserTest, which will test that typing
"remark" returns an instance of RemarkCommand.